// src/graph/types.ts
export type Maybe<T> = T | null;

export interface SearchIntent {
  id: Maybe<number>;

  agentId: Maybe<number>;

  isTopLevel: Maybe<boolean>;
}

export interface SearchIntentTrigger {
  id: Maybe<number>;

  value: Maybe<string>;

  type: Maybe<string>;
}

export interface IntentExpression {
  intentExpression: Maybe<string>;

  agentId: Maybe<number>;

  isTopLevel: Maybe<boolean>;
}